refactor: Move OrgStaffFactory helper to common/
